Supergrassi.jl

This is the reference documentation of Supergrassi.jl.

Index

Supergrassi.FilePath

Supergrassi.check_file_availability

Supergrassi.classify_dict

Supergrassi.clean_1d_values

Supergrassi.clean_2d_values

Supergrassi.clean_data

Supergrassi.clean_exports

Supergrassi.clean_household

Supergrassi.convert_to_ratio!

Supergrassi.convert_to_ratio!

Supergrassi.convert_to_ratio!

Supergrassi.correct_exports_with_services!

Supergrassi.create_filepath

Supergrassi.create_filepath_from_template

Supergrassi.read_data

Supergrassi.read_data

Supergrassi.rescale_data!

Supergrassi.round_shares!

Supergrassi.round_shares!

Supergrassi.round_shares!

Types

Supergrassi.FilePath — Type

FilePath(path::String)

A structure for file paths.

Fields

path::String: The full path.

directory::String: The directory.

file::String: The file name.

extension::String: The file extension.

Functions

Supergrassi.check_file_availability — Method

check_file_availability(settings::Dict)

Verifies the availability of data files.

Arguments

settings::Dict{<:Any, <:Any}: The settings dictionary.

Returns

Dict{String, FilePath}: A dictionary of file paths.

Examples

settings = Dict(
    "files" => Dict(
        "input_dir" => "data",
        "input_derived_dir" => "derived",
        "inputs" => Dict(
            "base" => Dict(
                "data" => "data.csv"
            ),
            "derived" => Dict(
                "scenario_independent" => Dict(
                    "data" => "data.csv"
                ),
                "scenario_dependent" => Dict(
                    "data" => "data.csv"
                )
            )
        )
    ),
    "version" => "v1",
    "files" => Dict(
        "scenario" => "1"
    )
)

filepaths = check_file_availability(settings)

Supergrassi.classify_dict — Method

classify_dict(dict::Dict)

Classifies a dictionary type based on the types of its keys and values. Used to create a typed dictionary from an untyped dictionary extracted from a YAML file.

Arguments

dict::Dict{Any, Any}: The dictionary to classify.

Returns

Dict{key_type, value_type}: The classified dictionary.

Examples

untyped_dict = YAML.load_file("settings.yaml")
typed_dict = classify_dict(dict)

Supergrassi.clean_1d_values — Method

Function to process the values that are split between uk, eu, world and stored in vectors.

Supergrassi.clean_2d_values — Method

Function to process the mValues stored in inputoutputmatrix:es in the data struct

Supergrassi.clean_data — Method

clean_data(data::Data, settings::Dict{String, Any})

Main function for data cleaning. Should take in a Data struct and return a CleanData struct.

Supergrassi.clean_exports — Method

Exports need a special treatment because they are a sum of export and services_export data frames. Service export is scaled by the sum of eu and world exports. Further refactoring definitely possible here.

Supergrassi.clean_household — Method

Function to process the household incomes and their derived data.

Supergrassi.convert_to_ratio! — Method

convert_to_ratio!(df::DataFrame)

Convert regional vector data into fractions of the aggregate value. Then renormalise the aggregate value by its sum. Nans are replaced with 0's.

Supergrassi.convert_to_ratio! — Method

convert_to_ratio!(data::InputMatrices)

Convert regional matrix data to fractions of the aggregate value Nans are replaced with 0's

Supergrassi.convert_to_ratio! — Method

convert_to_ratio!(data::RegionalData)

Convert regional data into ratios of region / sum(regions)

Supergrassi.correct_exports_with_services! — Method

Helper function for exports

Supergrassi.create_filepath — Method

create_filepath(basepath::String)

Creates a file path.

Arguments

file_path::FilePath: The file path object.

Returns

Bool: Whether the file path is valid.

Examples

file_path = FilePath("data.csv")
println(is_valid) # true

Supergrassi.create_filepath_from_template — Method

create_filepath_from_template(basepath::String, substitution_dict::Dict{String, Any})

Creates a file path from a template string and a dictionary of substitutions.

Arguments

basepath::String: The template string.

substitution_dict::Dict{String, Any}: The dictionary of substitutions.

Returns

FilePath: The file path object.

Examples

substitutions = Dict(
    "name" => "data",
    "format" => "csv",
    "directory" => "dir",
    "suffix" => 1,
)

template = "{directory}/{name}{suffix}.{format}"
filepath = create_filepath_from_template(template, substitutions)
println(filepath.path) # "dir/data1.csv"
println(filepath.directory) # "dir"
println(filepath.file) # "data1.csv"
println(filepath.extension) # ".csv"

Supergrassi.read_data — Method

read_data(file::String, sheet::String, range::String)

Reads a data file and returns a DataFrame.

Arguments

file::String: The path to the file.

sheet::String: The name of the sheet.

range::String: The range of cells to read.

Examples

df = read_data("data.xlsx", "Sheet1", "A1:Z1000")

Supergrassi.read_data — Method

read_data(file::String)

Reads a data file and returns a DataFrame.

Arguments

file::String: The path to the file.

Examples

df = read_data("data.csv")

Supergrassi.rescale_data! — Method

rescale_data!(data::CleanData)

Re-scale the data that does not get convereted into a ratio explicitly, following https://github.com/UCL/Supergrassi/blob/main/code/matlab/macro_v2/DataCleaning/RescaleData.m.

Supergrassi.round_shares! — Function

round_shares!(data::RegionalData, threshold::Float64 = 1e-4)

Round values below threshold in regional data to 0, then rescale so that regions sum to 1.

Supergrassi.round_shares! — Function

round_shares!(df::DataFrame, threshold = 1e-4)

Round shares in regional vector data

Supergrassi.round_shares! — Function

round_shares!(data::InputMatrices, threshold = 1e-4)

Round shares in regional matrix data

Constants