Supergrassi.jl

This is the reference documentation of Supergrassi.jl.

Index

Supergrassi.FilePath

Supergrassi.check_file_availability

Supergrassi.classify_dict

Supergrassi.clean_1d_values

Supergrassi.clean_2d_values

Supergrassi.clean_data

Supergrassi.clean_exports

Supergrassi.clean_household

Supergrassi.compute_all_parameters

Supergrassi.compute_objective_function

Supergrassi.compute_objective_function

Supergrassi.compute_parameter

Supergrassi.compute_production_parameter

Supergrassi.convert_to_ratio!

Supergrassi.convert_to_ratio!

Supergrassi.convert_to_ratio!

Supergrassi.correct_exports_with_services!

Supergrassi.create_filepath

Supergrassi.create_filepath_from_template

Supergrassi.firms_parameters_by_region

Supergrassi.log_eu_expenditure_on_uk_exports

Supergrassi.log_parameter_by_region

Supergrassi.log_parameters_by_region

Supergrassi.log_price_index

Supergrassi.log_total_price_index

Supergrassi.parameter_by_region

Supergrassi.parameters_by_region

Supergrassi.price_index

Supergrassi.productivity_shock_mean

Supergrassi.read_settings

Supergrassi.rescale_data!

Supergrassi.round_shares!

Supergrassi.round_shares!

Supergrassi.round_shares!

Supergrassi.total_capital_parameters

Supergrassi.total_input_parameters

Supergrassi.total_labor_parameters

Supergrassi.total_parameters

Types

Supergrassi.FilePath — Type

FilePath(path::String)

A structure for file paths.

Fields

path::String: The full path.

directory::String: The directory.

file::String: The file name.

extension::String: The file extension.

Functions

Supergrassi.check_file_availability — Method

check_file_availability(settings::Dict)

Verifies the availability of data files.

Arguments

settings::Dict{<:Any, <:Any}: The settings dictionary.

Returns

Dict{String, FilePath}: A dictionary of file paths.

Examples

settings = Dict(
    "files" => Dict(
        "input_dir" => "data",
        "input_derived_dir" => "derived",
        "inputs" => Dict(
            "base" => Dict(
                "data" => "data.csv"
            ),
            "derived" => Dict(
                "scenario_independent" => Dict(
                    "data" => "data.csv"
                ),
                "scenario_dependent" => Dict(
                    "data" => "data.csv"
                )
            )
        )
    ),
    "version" => "v1",
    "files" => Dict(
        "scenario" => "1"
    )
)

filepaths = check_file_availability(settings)

Supergrassi.classify_dict — Method

classify_dict(dict::Dict)

Classifies a dictionary type based on the types of its keys and values. Used to create a typed dictionary from an untyped dictionary extracted from a YAML file.

Arguments

dict::Dict{Any, Any}: The dictionary to classify.

Returns

Dict{key_type, value_type}: The classified dictionary.

Examples

untyped_dict = YAML.load_file("settings.yaml")
typed_dict = classify_dict(dict)

Supergrassi.clean_1d_values — Method

Function to process the values that are split between uk, eu, world and stored in vectors.

Supergrassi.clean_2d_values — Method

Function to process the mValues stored in inputoutputmatrix:es in the data struct

Supergrassi.clean_data — Method

clean_data(data::Data, settings::Dict{String, Any})

Main function for data cleaning. Should take in a Data struct and return a CleanData struct.

Supergrassi.clean_exports — Method

Exports need a special treatment because they are a sum of export and services_export data frames. Service export is scaled by the sum of eu and world exports. Further refactoring definitely possible here.

Supergrassi.clean_household — Method

Function to process the household incomes and their derived data.

Supergrassi.compute_all_parameters — Function

Compute all utility function parameters from regional data, elasticities and prices. Currently missing some of the γ parameters

fun parameter should be either parametersbyregion or logparametersby_region

Supergrassi.compute_objective_function — Method

compute_objective_function(x::Vector{<:Number}, data::CleanData, params::Parameters)

Computes the objective function value based on a vector of parameters.

Arguments

x::Vector{<:Number}: Vector containing log prices and zOC values.

data::CleanData: Cleaned data structure containing industry and regional information.

params::Parameters: Parameters structure containing production and constants.

Returns

objective_value::Float64: The computed objective function value.

Supergrassi.compute_objective_function — Method

function compute_objective_function(log_price_uk::Vector{<:Number}, zOC::Vector{<:Number}, data::CleanData, params::Parameters)

Computes the objective function value based on the log prices and zOC values.

Arguments

log_price_uk::Vector{<:Number}: Logarithm of UK prices.

zOC::Vector{<:Number}: Vector of zOC values.

data::CleanData: Cleaned data structure containing industry and regional information.

params::Parameters: Parameters structure containing production and constants.

Returns

objective_value::Float64: The computed objective function value.

Supergrassi.compute_parameter — Method

Compute 1d utility function parameters from a regional demand data frame and the corresponding elasticity. Currently missing the tilde parameters for exports.

Supergrassi.compute_production_parameter — Method

Compute the 2d utility function parameters from regional InputMatrices and the corresponding elasticity. Currently missing the jacobian, and the 1d parameters gammaL and gammaH

Supergrassi.convert_to_ratio! — Method

convert_to_ratio!(df::DataFrame)

Convert regional vector data into fractions of the aggregate value. Then renormalise the aggregate value by its sum. Nans are replaced with 0's.

Supergrassi.convert_to_ratio! — Method

convert_to_ratio!(data::InputMatrices)

Convert regional matrix data to fractions of the aggregate value Nans are replaced with 0's

Supergrassi.convert_to_ratio! — Method

convert_to_ratio!(data::RegionalData)

Convert regional data into ratios of region / sum(regions)

Supergrassi.correct_exports_with_services! — Method

Helper function for exports

Supergrassi.create_filepath — Method

create_filepath(basepath::String)

Creates a file path.

Arguments

file_path::FilePath: The file path object.

Returns

Bool: Whether the file path is valid.

Examples

file_path = FilePath("data.csv")
println(is_valid) # true

Supergrassi.create_filepath_from_template — Method

create_filepath_from_template(basepath::String, substitution_dict::Dict{String, Any})

Creates a file path from a template string and a dictionary of substitutions.

Arguments

basepath::String: The template string.

substitution_dict::Dict{String, Any}: The dictionary of substitutions.

Returns

FilePath: The file path object.

Examples

substitutions = Dict(
    "name" => "data",
    "format" => "csv",
    "directory" => "dir",
    "suffix" => 1,
)

template = "{directory}/{name}{suffix}.{format}"
filepath = create_filepath_from_template(template, substitutions)
println(filepath.path) # "dir/data1.csv"
println(filepath.directory) # "dir"
println(filepath.file) # "data1.csv"
println(filepath.extension) # ".csv"

Supergrassi.firms_parameters_by_region — Method

Compute utility function parameters (γ) by region considered in the model for the firms production function. This is a wrapper around parametersbyregion, but here we have to consider inputs to each firm from each firm which increases the dimension of the parameter array.

Matlab code reference ComputeTheta.m lines 257-259

Supergrassi.log_eu_expenditure_on_uk_exports — Method

Compute the share of EU expenditure on UK exports, defined in section 2.2.1 of the main paper.

Supergrassi.log_parameter_by_region — Method

Compute log of an utility function parameter for a single region

Supergrassi.log_parameters_by_region — Method

Compute log of utility function parameters. See parametersbyregion.

Supergrassi.log_price_index — Method

TODO

Supergrassi.log_total_price_index — Method

Compute the consumer price index defined in equation 2.7 of the main paper as ar{P}. Matlab code reference e.g. ComputeTheta.m line 58.

Supergrassi.parameter_by_region — Method

Compute an utility function parameter for a single region

Supergrassi.parameters_by_region — Method

Compute utility function parameters by region considered in the model (uk, eu, rest of world). Parameters of this type appear in multiple utility functions in the paper, and are annotated (at least) α, β1, β2, γ and ρ.

This is refactored from the Matlab code in e.g. ComputeTheta lines 62-64

Arguments

elasticity [ϵ, χ1, χ2, η, ξ]

logprice{uk, eu, w} [log(p)]

quantity_{uk, eu, w} [f, x1, x2, I, m]

Outputs

weight_{uk, eu, w}

Supergrassi.price_index — Method

Compute the the consumer price index defined in equation 2.3 of the main paper as ar{p}. Matlab code reference e.g. ComputeTheta.m line 49.

Supergrassi.productivity_shock_mean — Method

Compute the productivity shock mean μ

Supergrassi.read_settings — Method

read_settings(file_path::FilePath)

Reads settings from a YAML file and classifies the dictionary.

Arguments

file_path::FilePath: The path to the YAML file containing settings.

Returns

Dict{key_type, value_type}: The classified settings dictionary.

Supergrassi.rescale_data! — Method

rescale_data!(data::CleanData)

Re-scale the data that does not get convereted into a ratio explicitly, following https://github.com/UCL/Supergrassi/blob/main/code/matlab/macro_v2/DataCleaning/RescaleData.m.

Supergrassi.round_shares! — Function

round_shares!(df::DataFrame, threshold = 1e-4)

Round shares in regional vector data

Supergrassi.round_shares! — Function

round_shares!(data::RegionalData, threshold::Float64 = 1e-4)

Round values below threshold in regional data to 0, then rescale so that regions sum to 1.

Supergrassi.round_shares! — Function

round_shares!(data::InputMatrices, threshold = 1e-4)

Round shares in regional matrix data

Supergrassi.total_capital_parameters — Method

Compute the total parameter (γK) for the firms capital utility function.

Matlab code reference ComputeTheta.m line 254

Supergrassi.total_input_parameters — Method

Compute the total parameter (γM) for the firms input utility function.

Matlab code reference ComputeTheta.m line 251

Supergrassi.total_labor_parameters — Method

Compute the total parameter (γH) for the firms labor utility function.

Matlab code reference ComputeTheta.m line 249

Supergrassi.total_parameters — Method

Compute the total utility function parameter. Parameters of this type appear in multiple utility function in the paper, and are annotated (at least) α, β1, β2, γ and ρ.

This is refactored from the Matlab code in e.g. ComputeTheta.m line 59

Arguments

logpriceindex: price index computed by logpriceindex()

quantity: total quantity [f, x1, x2, I, m]

elasticity: [ϵ, χ1, χ2, η, ξ]

Outputs

parameters

Constants