Supergrassi.jl

This is the reference documentation of Supergrassi.jl.

Index

Supergrassi.AssetsLiabilities

Supergrassi.CleanData

Supergrassi.Constants

Supergrassi.Data

Supergrassi.Elasticities

Supergrassi.Elasticity

Supergrassi.ExchangeRates

Supergrassi.FilePath

Supergrassi.ForeignRegionalValues

Supergrassi.HouseholdData

Supergrassi.IndustryData

Supergrassi.InputMatrices

Supergrassi.InputOutput

Supergrassi.ParameterConstants

Supergrassi.Parameters

Supergrassi.ParamsProduction

Supergrassi.ParamsStruct

Supergrassi.RawHouseHoldData

Supergrassi.RawIncomeData

Supergrassi.RawIndustryData

Supergrassi.RegionalData

Supergrassi.Totals

Supergrassi.add_aggregate!

Supergrassi.capital_fun

Supergrassi.check_file_availability

Supergrassi.classify_dict

Supergrassi.clean_1d_values

Supergrassi.clean_2d_values

Supergrassi.clean_assets_liabilities

Supergrassi.clean_data

Supergrassi.clean_exports

Supergrassi.clean_household

Supergrassi.clean_matrix

Supergrassi.clean_rows

Supergrassi.clean_sigma_bar

Supergrassi.clean_vector

Supergrassi.combine_dataframe_row_wise

Supergrassi.compute_advalorem_tax

Supergrassi.compute_all_parameters

Supergrassi.compute_foreign_share

Supergrassi.compute_imports_shares

Supergrassi.compute_input_by_skill

Supergrassi.compute_objective_function

Supergrassi.compute_objective_function

Supergrassi.compute_parameter

Supergrassi.compute_production_parameter

Supergrassi.compute_wage_index

Supergrassi.convert_to_ratio!

Supergrassi.convert_to_ratio!

Supergrassi.convert_to_ratio!

Supergrassi.convert_to_ratio!

Supergrassi.correct_exports_with_services!

Supergrassi.create_filepath

Supergrassi.create_filepath_from_template

Supergrassi.create_map_105_to_64

Supergrassi.create_map_64_to_16

Supergrassi.generate_constants

Supergrassi.group_columns_by_new_name

Supergrassi.group_dataframes

Supergrassi.labor_fun

Supergrassi.log_eu_expenditure_on_uk_exports

Supergrassi.log_parameter_by_region

Supergrassi.log_parameters_by_region

Supergrassi.log_price_index

Supergrassi.log_total_price_index

Supergrassi.log_weight_kernel

Supergrassi.merge_quarterly_data

Supergrassi.parameter_by_region

Supergrassi.parameters_by_region

Supergrassi.parse_string_dataframe!

Supergrassi.postprocess_clean_data!

Supergrassi.price_index

Supergrassi.productivity_shock_mean

Supergrassi.read_csv

Supergrassi.read_excel

Supergrassi.read_settings

Supergrassi.reduce_columns_by_group_sum

Supergrassi.reduce_columns_by_group_weighted_mean

Supergrassi.rescale_data!

Supergrassi.round_shares!

Supergrassi.round_shares!

Supergrassi.round_shares!

Supergrassi.select_year

Supergrassi.sum_kernel

Supergrassi.total_capital_parameters

Supergrassi.total_input_parameters

Supergrassi.total_labor_parameters

Supergrassi.total_parameters

Supergrassi.weight_kernel

Types

Supergrassi.AssetsLiabilities — Type

struct AssetsLiabilities

In the matlab code these are called assets0 and assets1

Supergrassi.CleanData — Type

struct CleanData

Top level structure for cleaned up input data.

Contains

HouseholdData

IndustryData

Constants

Supergrassi.Constants — Type

struct Constants

Stores constants read from settings file.

Supergrassi.Data — Type

struct Data

Stores all the data required for the model, including household, industry, input-output data, and other relevant information.

Contains:

household: An instance of RawHouseHoldData containing household income and hours worked data.

industry: An instance of RawIndustryData containing industry capital, turnover, and inventory data.

input_output: An instance of InputOutput containing input-output data and matrices.

imports: An instance of InputOutput containing import data and matrices.

depreciation: A DataFrame containing depreciation data.

risk_free_rate: A DataFrame containing risk-free rate data.

assets: A DataFrame containing asset data.

model_results: A DataFrame containing model results.

merge_codes_105: A DataFrame containing merge codes for 105 industries.

merge_codes_64: A DataFrame containing merge codes for 64 industries.

others: A DataFrame containing other relevant data such as taxes and subsidies.

Supergrassi.Elasticities — Type

struct Elasticities

Stores sets of elasticity constants.

Names for reference in paper/matlab code:

consumption: α

eu_export_demand: β1

world_export_demand: β2

investment: ρ

production: ξ

Supergrassi.Elasticity — Type

struct Elasticity

Stores components of an elasticity constant.

Names for reference in paper/matlab code for e.g. production elasticity ξ:

substitution::Float64: ξ

armington::Float64: ξ_a

substitution_uk_other::Float64: ~ξ

skill_substitution::Float64: ξ_h

Supergrassi.ExchangeRates — Type

struct ExchangeRates

Stores exchange rates from GBP to other currencies

Supergrassi.FilePath — Type

FilePath(path::String)

A structure for file paths.

Fields

path::String: The full path.

directory::String: The directory.

file::String: The file name.

extension::String: The file extension.

Supergrassi.ForeignRegionalValues — Type

struct ForeignRegionalValues

Stores eu and rest of world components of a scalar quantity.

Supergrassi.HouseholdData — Type

struct HouseholdData

Stores data on households

Names reference for matlab code/paper

income::DataFrame: data.income_{lo, hi}

payments::DataFrame: data.{hValueLO, hValueHI, hValue}

hours::DataFrame: data.h{LO, HI}

wages::DataFrame: data.w{LO, HI}

Supergrassi.IndustryData — Type

struct IndustryData

Stores data on industries

Names reference for matlab code/paper

depreciation::DataFrame: data.depreciation

tax::DataFrame: data.taxValue{1,2}

capital::DataFrame: data.k{0,1}

surplus::DataFrame: data.kValue

shock_stdev::DataFrame: data.sigmaBar

assets_liabilities::AssetsLiabilities: AssetsLiabilities

regional::RegionalData: RegionalData

Supergrassi.InputMatrices — Type

struct InputMatrices

Stores intermediate input production matrices .

In the matlab code these matrices are called mValue_*

Supergrassi.InputOutput — Type

struct InputOutput

Stores input-output data matrix and other related information.

Contains:

raw_data: The raw input-output data as a DataFrame.

input_output_matrix: The input-output matrix as a DataFrame.

industry_names: An array of industry names.

final_consumption: An array of final consumption values.

gross_fixed_capital_formation: An array of gross fixed capital formation values.

delta_v_value_uk: An array of delta V values for the UK.

exports_eu_to_uk: An array of exports from the EU to the UK.

export_world_to_uk: An array of exports from the world to the UK.

total_use: An array of total use values.

services_export: An array of services export values.

Supergrassi.ParameterConstants — Type

struct ParameterConstants

Stores constant values that are stored in parameters but not updated

Supergrassi.Parameters — Type

struct Parameters

Main structure for parameters

Contains

constants::ParameterConstants : constant values

consumption::ParamsStruct : α

export_eu::ParamsStruct : β1

export_world::ParamsStruct : β2

production::ParamsProduction : γ

investment::ParamsStruct : ρ

Note: The constructor checks that all reasonable parameters are non-negative unless log or derivatives is set to true.

Supergrassi.ParamsProduction — Type

struct ParamsProduction

Stores firms intermediate production parameters

These require a different structure from ParamsStruct because firms trade with all other firms and the parameter is a matrix that contains a value for each combination of firms.

Names reference

input_human::Vector{Float64} : γ_h

input_capital::Vector{Float64} : γ_k

input_low_skill::Vector{Float64} : γ_L

input_high_skill::Vector{Float64} : γ_H

shock_mean::Vector{Float64} : μ

shock_stddev::Vector{Float64} : ̄σ

input_uk::Matrix{Float64} : γ_Md

input_eu::Matrix{Float64} : γ_Meu

input_world::Matrix{Float64} : γ_Mw

input_agg::Matrix{Float64} : γ_M

Supergrassi.ParamsStruct — Type

struct ParamsStruct

Stores parameters that are split between uk/eu/rest of the world.

Contains an optional field "tilde" which, for the export parameters β, is the share of foreign expenditure on UK exports.

Supergrassi.RawHouseHoldData — Type

struct RawHouseHoldData

Stores household data (income and hours worked).

Supergrassi.RawIncomeData — Type

struct RawIncomeData

Stores high and low income data as DataFrames.

Supergrassi.RawIndustryData — Type

struct RawIndustryData

Stores raw industry data (capital, turnover, and inventory) as DataFrames.

Supergrassi.RegionalData — Type

struct RegionalData

Stores industries data that is split between uk/eu/rest of the world.

The data is vectors with one value per industry stored in a DataFrame. Names for reference with the matlab code/paper:

total_use::DataFrame: data.yValue / y

consumption::DataFrame:data.fValue / f

delta_v::DataFrame: data.deltaVValue / Δv

export_eu::DataFrame: data.x1Value / x1

export_world::DataFrame: data.x2Value / x2

investment::DataFrame: data.IValue / I

input_matrices::InputMatrices: data.mValue / m InputMatrices

totals::Totals: data.{E, ISum, EX1, EX2} Totals

Supergrassi.Totals — Type

struct Totals

Stores sums of quantities over industries before renormalisation.

Names for reference with the paper/matlab code:

savings::Float64: E

investments::Float64: ISum

imports::ForeignRegionalValues: EX1, EX2 ForeignRegionalValues

Functions

Supergrassi.add_aggregate! — Method

function add_aggregate!(df::DataFrame)

Add an aggregate column to a DataFrame that sums the UK, EU, and World columns.

Arguments

df::DataFrame: The DataFrame to which the aggregate column will be added.

Supergrassi.capital_fun — Method

capital_fun(surplus::T, tau::T, output::T, capital::T, elasticity::T) where {T <: Real}

Argument:

surplus: Surplus value.

tau: Tax parameter.

output: total household use, data.yValue.

capital: firms capital.

elasticity: substitution elasticity parameter.

Supergrassi.check_file_availability — Method

check_file_availability(settings::Dict)

Verifies the availability of data files.

Arguments

settings::Dict{<:Any, <:Any}: The settings dictionary.

Returns

Dict{String, FilePath}: A dictionary of file paths.

Examples

settings = Dict(
    "files" => Dict(
        "input_dir" => "data",
        "input_derived_dir" => "derived",
        "inputs" => Dict(
            "base" => Dict(
                "data" => "data.csv"
            ),
            "derived" => Dict(
                "scenario_independent" => Dict(
                    "data" => "data.csv"
                ),
                "scenario_dependent" => Dict(
                    "data" => "data.csv"
                )
            )
        )
    ),
    "version" => "v1",
    "files" => Dict(
        "scenario" => "1"
    )
)

filepaths = check_file_availability(settings)

Supergrassi.classify_dict — Method

classify_dict(dict::Dict)

Classifies a dictionary type based on the types of its keys and values. Used to create a typed dictionary from an untyped dictionary extracted from a YAML file.

Arguments

dict::Dict{Any, Any}: The dictionary to classify.

Returns

Dict{key_type, value_type}: The classified dictionary.

Examples

untyped_dict = YAML.load_file("settings.yaml")
typed_dict = classify_dict(dict)

Supergrassi.clean_1d_values — Method

function clean_1d_values(val::Vector{<:Number}, val_imp::Vector{<:Number}, map_64::Dict{String, String}, map_16::Dict{String, String}, names_105::Vector, names_16::Vector, split_factor::Float64, industries_in_cols::Bool)

Function to process the values that are split between uk, eu, world and stored in vectors.

Arguments

val::Vector{<:Number}: The vector containing the values for UK.

val_imp::Vector{<:Number}: The vector containing the values for imports.

map_64::Dict{String, String}: A dictionary mapping SIC 64 industry names to SIC 16 industry names.

map_16::Dict{String, String}: A dictionary mapping SIC 16 industry names to their final names.

names_105::Vector: A vector of industry names for SIC 105.

names_16::Vector: A vector of industry names for SIC 16.

split_factor::Float64: The factor by which to split the imports between EU and World.

industries_in_cols::Bool: If true, industries will be on columns; otherwise, they will be on rows.

Returns

DataFrame: A DataFrame containing the cleaned values for UK, EU, World, and aggregate values.

Supergrassi.clean_2d_values — Method

function clean_2d_values(data::Data, split_factor::Float64)

Function to process the 2D values that are split between uk, eu, world and stored in matrices.

Arguments

data::Data: The Data struct containing input-output matrices and imports.

split_factor::Float64: The factor by which to split the imports between EU and World.

Returns

InputMatrices: A struct containing the cleaned input-output matrices for UK, EU, World, and aggregate values.

Supergrassi.clean_assets_liabilities — Function

funvtion clean_assets_liabilities(assets::DataFrame, year::Int64, map_to_16::Dict{String, String}, n_samples::Int64 = nrow(assets))

Wrapper function to clean assets and liabilities data.

Arguments

assets::DataFrame: The DataFrame containing assets and liabilities data.

year::Int64: The year for which the data is to be cleaned.

map_to_16::Dict{String, String}: A dictionary mapping SIC 64 industry names to SIC 16 industry names.

n_samples::Int64: The number of samples to limit per SIC 64 industry (default is the number of rows in assets.).

Returns

DataFrame: A cleaned DataFrame containing assets, liabilities, and their ratio, limited to n_samples per SIC 64 industry.

Supergrassi.clean_data — Method

clean_data(data::Data, settings::Dict{String, Any})

Main function for data cleaning. Should take in a Data struct and return a CleanData struct.

Arguments

data::Data: The Data struct containing various economic data.

settings::Dict{String, Any}: A dictionary containing constants and settings for the model.

Supergrassi.clean_exports — Method

function clean_exports(input_output::InputOutput, imports::InputOutput, split::Float64, names_16::Vector, industries_in_cols::Bool, map_64::Dict{String, String}, map_16::Dict{String, String})

Wrapper function to clean exports data.

Exports need a special treatment because they are a sum of export and services_export data frames. Service export is scaled by the sum of eu and world exports. Further refactoring definitely possible here.

Arguments

input_output::InputOutput: The InputOutput struct containing export data.

imports::InputOutput: The InputOutput struct containing import data.

split::Float64: The factor by which to split the imports between EU and World.

names_16::Vector: A vector of industry names for SIC 16.

industries_in_cols::Bool: If true, industries will be on columns; otherwise, they will be on rows.

map_64::Dict{String, String}: A dictionary mapping SIC 64 industry names to SIC 16 industry names.

map_16::Dict{String, String}: A dictionary mapping SIC 16 industry names to their final names.

Returns

DataFrame, DataFrame: Two DataFrames containing exports to the EU and World, respectively.

Supergrassi.clean_household — Method

function clean_household(data::Data, year::Int64, map_64::Dict{String, String}, map_16::Dict{String, String}, names_105::Vector, names_64::Vector, names_16::Vector, industries_in_cols::Bool)

Function to process the household incomes and their derived data.

Arguments

data::Data: The Data struct containing household income and hours data.

year::Int64: The year for which the data is to be cleaned.

map_64::Dict{String, String}: A dictionary mapping SIC 105 industry names to SIC 64 industry names.

map_16::Dict{String, String}: A dictionary mapping SIC 64 industry names to SIC 16 industry names.

names_105::Vector: A vector of industry names for SIC 105.

names_64::Vector: A vector of industry names for SIC 64.

names_16::Vector: A vector of industry names for SIC 16.

industries_in_cols::Bool: If true, industries will be on columns; otherwise, they will be on rows.

Returns

HouseholdData: A struct containing cleaned household income, payments, hours, and wages data.

Supergrassi.clean_matrix — Method

function clean_matrix(data::DataFrame, industry_names::Array{String, 1}, mapping::Dict{String, String})

Clean a matrix of data by reducing based on a mapping and renaming them with industry names.

Arguments

data::DataFrame: A DataFrame containing the matrix data to be cleaned.

industry_names::Array{String, 1}: An array of industry names to be used as column names.

mapping::Dict{String, String}: A dictionary mapping old column names to new column names for reduction.

Returns

DataFrame: A new DataFrame with the matrix data, renamed columns, and reduced columns based on the mapping.

Supergrassi.clean_rows — Method

function clean_rows(data::DataFrame, row_names::String, col_names::Array{String, 1}, mapping::Dict{String, String})

Clean selected subset of rows in a DataFrame based on row names and rename columns.

Arguments

data::DataFrame: The DataFrame containing the data to be cleaned.

row_names::String: The name of the row to select.

col_names::Array{String, 1}: An array of new column names to rename the selected columns.

mapping::Dict{String, String}: A dictionary mapping old column names to new column names for reduction.

Returns

DataFrame: A new DataFrame with selected rows, renamed columns, and reduced columns based on the mapping.

Supergrassi.clean_sigma_bar — Method

clean_sigma_bar(sigma_data::Vector, zero_list::Vector{Int64}, sic64::Vector{String})

Clean the sigma bar data by converting it to a DataFrame, parsing strings to Float64, and setting specified indices to zero.

Arguments

sigma_data::Vector: A vector containing sigma bar data.

zero_list::Vector{Int64}: A vector of indices where the sigma bar values should be set to zero.

sic64::Vector{String}: A vector of SIC 64 industry names to be used as column names.

Returns

DataFrame: A DataFrame containing the cleaned sigma bar data, with specified indices set to zero.

Supergrassi.clean_vector — Method

function clean_vector(data::Vector{<:Number}, industry_names::Array{String, 1}, mapping::Dict{String, String})

Clean a vector of numbers by creating a DataFrame, renaming columns, and reducing columns based on a mapping.

Arguments

data::Vector{<:Number}: A vector of numbers to be cleaned.

industry_names::Array{String, 1}: An array of industry names to be used as column names.

mapping::Dict{String, String}: A dictionary mapping old column names to new column names for reduction.

Returns

DataFrame: A new DataFrame with the vector data, renamed columns, and reduced columns based on the mapping.

Supergrassi.combine_dataframe_row_wise — Method

function combine_dataframe_row_wise(data::DataFrame, func::Function)

Combine rows of a DataFrame by applying a function to each row.

Arguments

data::DataFrame: The DataFrame containing the data to be combined.

func::Function: A function to apply to each row of the DataFrame.

Returns

DataFrame: A new DataFrame with the combined results, where each column is named after the original column names.

Supergrassi.compute_advalorem_tax — Method

compute_advalorem_tax(data::IndustryData)

Compute the ad valorem tax rate by combining product and production taxes, normalized by total use.

Arguments

data::IndustryData: Industry data containing tax and regional total use information

Supergrassi.compute_all_parameters — Function

function compute_all_parameters(data::CleanData, prices::DataFrame, fun::Function = parameters_by_region)

Compute all utility function parameters and their derivativesfrom regional data, elasticities and prices.

For description of the parameters, see Table 2 of Nesheim et al (in prep). Return two Parameters structs, the first one contains values of parameters α, β1, β2, ρ and γ, the second one contains their derivatives with respect to prices.uk.

See the tests in testparameterswith_data.jl for an example of use.

Arguments

data::CleanData: Data structure created by clean_data

prices::DataFrame: Logarithm of price index equilibrium variable. Has three components (uk, eu, world)

fun::Function=parameters_by_region: Function that computes three parameters by region. Either parameters_by_region or log_parameters_by_region

See also compute_parameter, compute_foreign_share, compute_production_parameter, Parameters

Supergrassi.compute_foreign_share — Method

compute_foreign_share(param::ParamsStruct, dparam::ParamsStruct, demand::DataFrame,
                      elasticity::Elasticity, prices::DataFrame, E::T, Ex::T,
                      PTilde::T, exchange_rate::T) where {T<:Real}

Compute the share of foreign expenditure on UK exports, β̃ in the paper by Nesheim et al.

Computes the value and derivative of β̃ and adds them to the structures param and dparam, respectively. Because the strucs are immutable, a copy is created. Requires input outside of the regional demand data which is why this has been split into a different function from compute_parameter.

Arguments

param::ParamsStruct: Parameters.

dparam::ParamsStruct: Derivatives of parameters.

demand::DataFrame: Demand data disaggregated by region.

elasticity::Elasticity: Values for elasticity of substitution

prices::DataFrame: Logarithm of price index equilibrium variable. Disagregated by region.

E::Real: Household expenditure

Ex::Real: Foreign expenditure on UK exports

PTilde::Real: The foreign price index.

exchange_rate::Real: The exchange rate between domestic and foreign currency.

Supergrassi.compute_imports_shares — Method

compute_imports_shares(constants::Constants)

Compute the share of UK imports in total imports for EU and world regions.

Arguments

constants::Constants: Constants containing import and exchange rate information

Supergrassi.compute_input_by_skill — Method

compute_input_by_skill(data::HouseholdData, elasticity::Elasticity)

Compute input shares by skill level (low and high skill) based on household payments and wages.

Arguments

data::HouseholdData: Household data containing payments and wages information

elasticity::Elasticity: Elasticity of substitution for production parameters

Supergrassi.compute_objective_function — Method

compute_objective_function(x::Vector{<:Number}, data::CleanData, params::Parameters)

Computes the objective function value based on a vector of parameters.

Arguments

x::Vector{<:Number}: Vector containing log prices and zOC values.

data::CleanData: Cleaned data structure containing industry and regional information.

params::Parameters: Parameters structure containing production and constants.

Returns

objective_value::Float64: The computed objective function value.

Supergrassi.compute_objective_function — Method

function compute_objective_function(log_price_uk::Vector{<:Number}, zOC::Vector{<:Number}, data::CleanData, params::Parameters)

Computes the objective function value based on the log prices and zOC values.

Arguments

log_price_uk::Vector{<:Number}: Logarithm of UK prices.

zOC::Vector{<:Number}: Vector of zOC values.

data::CleanData: Cleaned data structure containing industry and regional information.

params::Parameters: Parameters structure containing production and constants.

Returns

objective_value::Float64: The computed objective function value.

Supergrassi.compute_parameter — Method

function compute_parameter(demand::DataFrame, elasticity::Elasticity, prices::DataFrame, fun::Function = parameters_by_region)

Compute 1d utility function parameters and their derivatives from a regional demand DataFrame and the corresponding elasticity.

Arguments

demand::DataFrame: Demand data disaggregated by region

elasticity::Elasticity: Values for elasticity of substitution

prices::DataFrame: Logarithm of price index equilibrium variable. Disagregated by region.

fun::Function = parameters_by_region: Function that computes three parameters by region. Either parameters_by_region or log_parameters_by_region

Supergrassi.compute_production_parameter — Method

function compute_production_parameter(data::CleanData, prices::DataFrame, fun::Function = parameters_by_region)

Compute the 2d utility function parameters γM, γH, γK from regional InputMatrices and the corresponding elasticity.

Arguments

data::CleanData: Data structure created by clean_data

prices::DataFrame: Logarithm of price index equilibrium variable. Has three components (uk, eu, world)

fun::Function=parameters_by_region: Function that computes three parameters by region. Either parameters_by_region or log_parameters_by_region

Supergrassi.compute_wage_index — Method

compute_wage_index(data::HouseholdData, elasticity::Elasticity)

Computes aggregate wage index for households based on household data and elasticity of production.

ref

Arguments

data::HouseholdData: Data structure containing wages and payments DataFrames HouseholdData

elasticity::Elasticity: Elasticity of substitution for production parameters (ξ)

Supergrassi.convert_to_ratio! — Method

convert_to_ratio!(df::DataFrame)

Convert regional vector data into fractions of the aggregate value. Then renormalise the aggregate value by its sum. Nans are replaced with 0's.

Arguments

df::DataFrame: The DataFrame containing regional vector data to be converted into ratios.

Supergrassi.convert_to_ratio! — Method

convert_to_ratio!(data::HouseholdData)

Convert household data into ratios of low and high payments to the sum of low and high payments.

Arguments

data::HouseholdData: The HouseholdData struct containing household data to be converted into ratios.

Supergrassi.convert_to_ratio! — Method

convert_to_ratio!(data::InputMatrices)

Convert regional matrix data to fractions of the aggregate value Nans are replaced with 0's

Arguments

data::InputMatrices: The InputMatrices struct containing regional matrix data to be converted into ratios.

Supergrassi.convert_to_ratio! — Method

convert_to_ratio!(data::RegionalData)

Convert regional data into ratios of region / sum(regions)

Arguments

data::RegionalData: The RegionalData struct containing regional data to be converted into ratios.

Supergrassi.correct_exports_with_services! — Method

function correct_exports_with_services!(export_to_eu::DataFrame, export_to_world::DataFrame, services_export::DataFrame)

Function to correct exports by accounting for NaN values and scaling appropriately.

Arguments

export_to_eu::DataFrame: DataFrame containing exports to the EU.

export_to_world::DataFrame: DataFrame containing exports to the World.

services_export::DataFrame: DataFrame containing service exports.

Supergrassi.create_filepath — Method

create_filepath(basepath::String)

Creates a file path.

Arguments

file_path::FilePath: The file path object.

Returns

Bool: Whether the file path is valid.

Examples

file_path = FilePath("data.csv")
println(is_valid) # true

Supergrassi.create_filepath_from_template — Method

create_filepath_from_template(basepath::String, substitution_dict::Dict{String, Any})

Creates a file path from a template string and a dictionary of substitutions.

Arguments

basepath::String: The template string.

substitution_dict::Dict{String, Any}: The dictionary of substitutions.

Returns

FilePath: The file path object.

Examples

substitutions = Dict(
    "name" => "data",
    "format" => "csv",
    "directory" => "dir",
    "suffix" => 1,
)

template = "{directory}/{name}{suffix}.{format}"
filepath = create_filepath_from_template(template, substitutions)
println(filepath.path) # "dir/data1.csv"
println(filepath.directory) # "dir"
println(filepath.file) # "data1.csv"
println(filepath.extension) # ".csv"

Supergrassi.create_map_105_to_64 — Function

function create_map_105_to_64(merge_codes::DataFrame, verbose::Bool = false)

Create a mapping from SIC 105 industry names to SIC 64 industry names.

Arguments

merge_codes::DataFrame: A DataFrame containing the mapping between SIC 105 and SIC 64 industry names. It should have two columns: sic105 containing SIC 105 industry names and sic64 containing SIC 64 industry names.

verbose::Bool: If true, prints the initial and final industry names.

Returns

Dict{String, String}: A dictionary mapping SIC 105 industry names to SIC 64 industry names.

Supergrassi.create_map_64_to_16 — Function

function create_map_105_to_64(merge_codes::DataFrame, verbose::Bool = false)

Create a mapping from SIC 64 industry names to SIC 16 industry names.

Arguments

merge_codes::DataFrame: A DataFrame containing the mapping between SIC 64 and SIC 16 industry names. It should have two columns: x1 containing SIC 64 industry names and x7 containing SIC 16 industry names.

verbose::Bool: If true, prints the initial and final industry names.

Returns

Dict{String, String}: A dictionary mapping SIC 64 industry names to SIC 16 industry names.

Supergrassi.generate_constants — Method

generate_constants(data::Data, settings::Dict{String, Any})

Generate a Constants struct from the provided data and settings.

Arguments

data::Data: The Data struct containing various economic data.

settings::Dict{String, Any}: A dictionary containing constants and settings for the model.

Returns

Constants: A Constants struct containing the year, exchange rates, interest rate, total imports, import tariffs, export costs, and elasticities.

Supergrassi.group_columns_by_new_name — Method

function group_columns_by_new_name(mapping::Dict{String, String})

Group columns by their new names based on a mapping.

Arguments

mapping::Dict{String, String}: A dictionary mapping old column names to new column names.

Returns

Dict{String, Vector{Symbol}}: A dictionary where keys are new column names and values are vectors of old column names (as Symbols) that map to them.

Supergrassi.group_dataframes — Function

function group_dataframes(dfs::AbstractArray, col_names::AbstractArray, industry_names::AbstractArray, industries_on_cols::Bool = true, reduction_fun::Function = nothing, mapping::Dict{String, String} = Dict(); kwargs...)

Group multiple DataFrames by industry names and aggregate their values.

Arguments

dfs::AbstractArray: An array of DataFrames to be grouped.

col_names::AbstractArray: An array of column names for the resulting DataFrame.

industry_names::AbstractArray: An array of industry names to be used as row names.

industries_on_cols::Bool: If true, industries will be on columns; otherwise, they will be on rows.

reduction_fun::Function: A function to apply for data aggregation (default is nothing).

mapping::Dict{String, String}: A dictionary mapping old column names to new column names.

kwargs...: Additional keyword arguments to be passed to the reduction_fun.

Returns

DataFrame: A DataFrame with industries as rows or columns, depending on industries_on_cols, and aggregated values.

Supergrassi.labor_fun — Method

labor_fun(labor::T, log_wages::T, elasticity::T) where {T <: Real}

Argument:

labor: Labor input.

log_wages: Log of wages.

elasticity: Elasticity parameter.

See also

compute_wage_index

Supergrassi.log_eu_expenditure_on_uk_exports — Method

function log_eu_expenditure_on_uk_exports(log_price_index::Vector{T}, quantity::Vector{T},Ex::T, ETilde::T, ePx::T, PTilde::T, elasticity::T,elasticity_tilde::T) where {T <: Real}

Compute the share of EU expenditure on UK exports, defined in section 2.2.1 of the main paper.

for the export parameters β, this is the share of foreign expenditure on UK exports.

Arguments

log_price_index::Vector{Real} : price index computed by log_price_index

quantity::Vector{Real} : Quantity of domestically-produced good i exported to the EU

Ex::Real : sum of imports

Etilde::Real: EU expenditure on UK exports

ePx::Real : exchange rate to foreign currency

Ptilde::Real : UK exportprice index

elasticity::Real : substitution elasticity

elasticity_tilde::Real : substitution from uk to other elasticity

See also ParamsStruct.

Supergrassi.log_parameter_by_region — Method

function log_parameter_by_region(elasticity::T, quantity_region::T, logP_region::T, logP::T) where {T <: Real}

Compute log of an utility function parameter for a single region. See parameter_by_region

Supergrassi.log_parameters_by_region — Method

function log_parameters_by_region(elasticity::T,log_price_uk::T,log_price_eu::T,log_price_world::T,quantity_uk::T,quantity_eu::T,quantity_world::T) where {T <: Real}

Compute log of utility function parameters. See parameters_by_region.

Supergrassi.log_price_index — Method

function log_price_index(elasticity::T,log_price_uk::T, log_price_eu::T, log_price_world::T,demand_uk::T, demand_eu::T, demand_world::T) where {T <: Real}

Compute the log of the consumer price index. See price_index

Supergrassi.log_total_price_index — Method

function log_total_price_index(elasticity::T, log_price_index::Vector{T}, quantity::Vector{T}) where {T <: Real}

Compute the total consumer price index defined in equation 2.7 of the main paper as ar{P}. Matlab code reference e.g. ComputeTheta.m line 58.

Arguments

elasticity::Real : substitution elasticity

log_price_index::Vector{Real} : price index computed by log_price_index

quantity:Vector{Real} : aggregate quantity [f, x1, x2, I, m]

Supergrassi.log_weight_kernel — Method

log_weight_kernel(a::T, b::T, elasticity::T) where {T <: Real}

computes log(a/b^(1-elasticity))

Argument:

a: Numerator value (in log form).

b: Denominator base value (in log form).

elasticity: substitution elasticity parameter

Supergrassi.merge_quarterly_data — Method

function merge_quarterly_data(df::DataFrame, year::Int64, industry_names::Vector, fun::Function)

Merge quarterly data for a specific year by applying a function to each row.

Arguments

df::DataFrame: The DataFrame containing quarterly data.

year::Int64: The year for which the data is to be merged.

industry_names::Vector: A vector of industry names to be used as row names.

fun::Function: A function to combine the quarterly data (e.g., sum).

Returns

DataFrame: A new DataFrame with the merged data for the specified year, with industry names as row names.

Supergrassi.parameter_by_region — Method

function parameter_by_region(elasticity::T, quantity_region::T, logP_region::T, logP::T) where {T <: Real}

Compute an utility function parameter for a single region

Arguments

elasticity::Real : substitution elasticity

quantitity_region::Real : quantity for the region. One of [f, x1, x2, I, m]

logP_region::Real : log of price index for region

logP::Real : log of aggregate price index

Supergrassi.parameters_by_region — Method

function parameters_by_region(elasticity::T,log_price_uk::T,log_price_eu::T,log_price_world::T,quantity_uk::T,quantity_eu::T,quantity_world::T) where {T <: Real}

Compute utility function parameters by region considered in the model (uk, eu, rest of world). Parameters of this type appear in multiple utility functions in the paper, and are annotated (at least) α, β1, β2, γ and ρ.

This is refactored from the Matlab code in e.g. ComputeTheta lines 62-64

Arguments

elasticity::Real : substitution elasticity. One of [ϵ, χ1, χ2, η, ξ]

log_price_{uk, eu, w}::Real : log price indices for each region [log(p)]

quantity_{uk, eu, w}::Real : quantity for the region. One of [f, x1, x2, I, m]

Supergrassi.parse_string_dataframe! — Function

function parse_string_dataframe!(df::DataFrame, T::Type, default_val=nothing)

Parse string columns in a DataFrame to a specified type, replacing missing values with a default value.

Arguments

df::DataFrame: The DataFrame containing the string columns to be parsed.

T::Type: The type to which the string columns should be parsed (e.g., Float64).

default_val: The value to replace missing values with (default is nothing).

Supergrassi.postprocess_clean_data! — Method

postprocess_clean_data!(data::CleanData)

Post-process the cleaned data to convert regional data into ratios, round shares, and rescale the data.

Arguments

data::CleanData: The CleanData struct containing household and industry data to be post-processed.

Supergrassi.price_index — Method

function price_index(elasticity::T,log_price_uk::T, log_price_eu::T, log_price_world::T,demand_uk::T, demand_eu::T, demand_world::T) where {T <: Real}

Compute the the consumer price index defined in equation 2.3 of the main paper as ar{p}. Matlab code reference e.g. ComputeTheta.m line 49.

Arguments

elasticity::Real : substitution elasticity

log_price_{uk, eu, w}::Real : log price indices for each region [log(p)]

demand_{uk, eu, w}::Real : demand for the region. One of [f, x1, x2, I, m]

Supergrassi.productivity_shock_mean — Method

function productivity_shock_mean(elasticity::Elasticity, prices_uk::Vector{T}, prices_eu::Vector{T}, prices_world::Vector{T},
                             input_uk::Vector{T}, input_eu::Vector{T}, input_world::Vector{T}, input_agg::Vector{T},
                             surplus::T, capital::T, output::T, labor::T, log_wages::T, tau::T, row::Int, log_scale::Bool) where {T <: Real}

Compute the productivity shock mean μ

prices_uk::Vector{Real} : uk price index

prices_eu::Vector{Real} : eu price index

prices_world::Vector{Real} : rest of world price index

input_uk::Vector{Real} : uk firms intermediate input

input_eu::Vector{Real} : eu firms intermediate input

input_world::Vector{Real} : rest of world firms intermediate input

input_agg::Vector{Real} : aggreagete firms intermediate input

suprlus::Real : firms suprlus, data.kValue

capital::Real : firms capital, data.k0

output::Real : total household use, data.yValue

labor::Real : payments for labor, data.hValue

log_wages::Real : aggregate wages of households

elasticity::Real : substitution elasticity

tau::Real : ratio of taxes to household use

row::Int : index of firm i

log_scale::Bool : return parameters in log scale

Supergrassi.read_csv — Method

read_csv(file::String)

Reads a CSV file and returns a DataFrame.

Arguments

file::String: The path to the CSV file.

Returns

DataFrame: A DataFrame containing the data from the CSV file.

Examples

df = read_csv("data.csv")

Supergrassi.read_excel — Method

read_excel(file::String; sheet::String="Sheet1", range::String="A1:Z1000")

Reads an Excel file and returns a DataFrame.

Arguments

file::String: The path to the Excel file.

sheet::String: The name of the sheet to read. Default is "Sheet1".

range::String: The range of cells to read, in Excel format (e.g., "A1:Z1000"). Default is "A1:Z1000".

Returns

DataFrame: A DataFrame containing the data from the specified sheet and range.

Examples

df = read_excel("data.xlsx", sheet="Sheet1", range="A1:Z1000")

Supergrassi.read_settings — Method

read_settings(file_path::FilePath)

Reads settings from a YAML file and classifies the dictionary.

Arguments

file_path::FilePath: The path to the YAML file containing settings.

Returns

Dict{key_type, value_type}: The classified settings dictionary.

Supergrassi.reduce_columns_by_group_sum — Method

function reduce_columns_by_group_sum(df::DataFrame, mapping::Dict{String, String})

Reduce columns in a DataFrame by summing them based on a mapping.

Arguments

df::DataFrame: The DataFrame containing the columns to be reduced.

mapping::Dict{String, String}: A dictionary mapping old column names to new column names.

Returns

DataFrame: A new DataFrame with reduced columns, where each new column is the sum of the old columns that map to it.

Supergrassi.reduce_columns_by_group_weighted_mean — Method

fubction reduce_columns_by_group_weighted_mean(df::DataFrame, mapping::Dict{String, String}; weights::DataFrame = DataFrame(ones(1,ncol(df)), names(df)))

Reduce columns in a DataFrame by calculating the weighted mean based on a mapping.

Arguments

df::DataFrame: The DataFrame containing the columns to be reduced.

mapping::Dict{String, String}: A dictionary mapping old column names to new column names.

weights::DataFrame: A DataFrame containing weights for each column. Defaults to a DataFrame of ones.

Returns

DataFrame: A new DataFrame with reduced columns, where each new column is the weighted mean of the old columns that map to it.

Supergrassi.rescale_data! — Method

rescale_data!(data::CleanData)

Re-scale the data that does not get convereted into a ratio explicitly, following https://github.com/UCL/Supergrassi/blob/main/code/matlab/macro_v2/DataCleaning/RescaleData.m.

Arguments

data::CleanData: The CleanData struct containing household and industry data to be rescaled.

Supergrassi.round_shares! — Function

round_shares!(df::DataFrame, threshold = 1e-4)

Round shares in regional vector data

Arguments

df::DataFrame: The DataFrame containing regional data to be rounded.

threshold: The threshold below which values are set to 0 (default is 1e-4).

Supergrassi.round_shares! — Function

round_shares!(data::InputMatrices, threshold = 1e-4)

Round shares in regional matrix data

Arguments

data::InputMatrices: The InputMatrices struct containing regional matrix data to be rounded.

Supergrassi.round_shares! — Function

round_shares!(data::RegionalData, threshold::Float64 = 1e-4)

Round values below threshold in regional data to 0, then rescale so that regions sum to 1.

Arguments

data::RegionalData: The RegionalData struct containing regional data to be rounded.

Supergrassi.select_year — Method

function select_year(data::DataFrame, year::Int64)

Select rows from a DataFrame based on a specific year.

Arguments

data::DataFrame: The DataFrame containing the data.

year::Int64: The year to filter the DataFrame by.

Returns

DataFrame: A new DataFrame containing only the rows for the specified year, with columns converted to Float64 if they are strings.

Supergrassi.sum_kernel — Method

sum_kernel(var::Vector{T}, logP::Vector{T}, elasticity::T) where {T <: Real}

Argument:

var: Vector of variable values.

logP: Vector of log prices.

elasticity: substitution elasticity parameter

Supergrassi.total_capital_parameters — Method

function total_capital_parameters(prices_uk::Vector{T}, prices_eu::Vector{T}, prices_world::Vector{T},
                              input_uk::Vector{T}, input_eu::Vector{T}, input_world::Vector{T}, input_agg::Vector{T},
                              surplus::T, capital::T, output::T, labor::T, log_wages::T, elasticity::Elasticity, tau::T, log_scale::Bool) where T

Compute the total parameter (γK) for the firms capital utility function.

Matlab code reference ComputeTheta.m line 254

Arguments

prices_uk::Vector{Real} : uk price index

prices_eu::Vector{Real} : eu price index

prices_world::Vector{Real} : rest of world price index

input_uk::Vector{Real} : uk firms intermediate input

input_eu::Vector{Real} : eu firms intermediate input

input_world::Vector{Real} : rest of world firms intermediate input

input_agg::Vector{Real} : aggreagete firms intermediate input

suprlus::Real : firms suprlus, data.kValue

capital::Real : firms capital, data.k0

output::Real : total household use, data.yValue

labor::Real : payments for labor, data.hValue

log_wages::Real : aggregate wages of households

elasticity::Real : substitution elasticity

tau::Real : ratio of taxes to household use

log_scale::Bool : return parameters in log scale

See also

compute_wage_index

Supergrassi.total_input_parameters — Method

function total_input_parameters(prices_uk::Vector{T}, prices_eu::Vector{T}, prices_world::Vector{T},
                            input_uk::Vector{T}, input_eu::Vector{T}, input_world::Vector{T}, input_agg::Vector{T},
                            surplus::T, capital::T, output::T, labor::T, log_wages::T, elasticity::Elasticity, tau::T, log_scale::Bool) where {T <: Real}

Compute the total parameter (γM) for the firms input utility function.

Matlab code reference ComputeTheta.m line 251

Arguments

prices_uk::Vector{Real} : uk price index

prices_eu::Vector{Real} : eu price index

prices_world::Vector{Real} : rest of world price index

input_uk::Vector{Real} : uk firms intermediate input

input_eu::Vector{Real} : eu firms intermediate input

input_world::Vector{Real} : rest of world firms intermediate input

input_agg::Vector{Real} : aggreagete firms intermediate input

suprlus::Real : firms suprlus, data.kValue

capital::Real : firms capital, data.k0

output::Real : total household use, data.yValue

labor::Real : payments for labor, data.hValue

log_wages::Real : aggregate wages of households

elasticity::Real : substitution elasticity

tau::Real : ratio of taxes to household use

log_scale::Bool : return parameters in log scale

See also

compute_wage_index

Supergrassi.total_labor_parameters — Method

function total_labor_parameters(prices_uk::Vector{T}, prices_eu::Vector{T}, prices_world::Vector{T},
                            input_uk::Vector{T}, input_eu::Vector{T}, input_world::Vector{T}, input_agg::Vector{T},
                            surplus::T, capital::T, output::T, labor::T, log_wages::T, elasticity::Elasticity, tau::T, log_scale::Bool) where T

Compute the total parameter (γH) for the firms labor utility function.

Matlab code reference ComputeTheta.m line 249

Arguments

prices_uk::Vector{Real} : uk price index

prices_eu::Vector{Real} : eu price index

prices_world::Vector{Real} : rest of world price index

input_uk::Vector{Real} : uk firms intermediate input

input_eu::Vector{Real} : eu firms intermediate input

input_world::Vector{Real} : rest of world firms intermediate input

input_agg::Vector{Real} : aggreagete firms intermediate input

suprlus::Real : firms suprlus, data.kValue

capital::Real : firms capital, data.k0

output::Real : total household use, data.yValue

labor::Real : payments for labor, data.hValue

log_wages::Real : aggregate wages of households

elasticity::Real : substitution elasticity

tau::Real : ratio of taxes to household use

log_scale::Bool : return parameters in log scale

See also

compute_wage_index

Supergrassi.total_parameters — Method

function total_parameters(log_price_index::Vector{T}, quantity::Vector{T}, elasticity::T ) where {T <: Real}

Compute the aggregate utility function parameter. Parameters of this type appear in multiple utility function in the paper, and are annotated (at least) α, β1, β2, γ and ρ.

This is refactored from the Matlab code in e.g. ComputeTheta.m line 59

Arguments

log_price_index::Vector{Real} : price index computed by log_price_index

quantity:Vector{Real} : aggregate quantity [f, x1, x2, I, m]

elasticity::Real : substitution elasticity. One of [ϵ, χ1, χ2, η, ξ]

Supergrassi.weight_kernel — Method

weight_kernel(a::T, b::T, elasticity::T) where {T <: Real}

Computes a/b^(1-elasticity)

Argument:

a: Numerator value.

b: Denominator value.

elasticity: substitution elasticity parameter

Constants